fex changes with sceletons and handleGoHome
