+++ components/docInIframe.tsx
@@ -1,7 +1,6 @@
 import { useEffect, useState } from "react"
-import IframeResizer from 'iframe-resizer-react';
-import SubmitButton from "./buttons/submitButton";
 import { ExternalLinkIcon } from "@heroicons/react/outline";
+import { DocInFrameSceleton } from "./Sceletons";
 
 type Props = {
     URl: string;
@@ -19,7 +18,7 @@ export function DocIframe({ URl, onConfirm }: Props) {
         <div className="text-white text-base scrollbar:!w-1.5 scrollbar:!h-1.5 scrollbar:bg-darkblue-500 scrollbar-track:!bg-slate-100 scrollbar-thumb:!rounded scrollbar-thumb:!bg-slate-300 scrollbar-track:!rounded scrollbar-track:!bg-slate-500/[0.16] scrollbar-thumb:!bg-slate-500/50">
             <div className={`relative ${loading ? '' : 'pb-96'} scrollbar:!w-1.5 scrollbar:!h-1.5 scrollbar:bg-darkblue-500 scrollbar-track:!bg-slate-100 scrollbar-thumb:!rounded scrollbar-thumb:!bg-slate-300 scrollbar-track:!rounded scrollbar-track:!bg-slate-500/[0.16] scrollbar-thumb:!bg-slate-500/50`}>
                 {
-                    loading && <Sceleton />
+                    loading && <DocInFrameSceleton />
                 }
                 <iframe src={URl} className={`${loading ? 'invisible h-0 w-0' : 'visible animate-fade-in-down'} scrollbar:!w-1.5 scrollbar:!h-1.5 scrollbar:bg-darkblue-500 scrollbar-track:!bg-slate-100 scrollbar-thumb:!rounded scrollbar-thumb:!bg-slate-300 scrollbar-track:!rounded scrollbar-track:!bg-slate-500/[0.16] scrollbar-thumb:!bg-slate-500/50 border-0 self-center absolute w-full h-full`}></iframe>
 
@@ -47,29 +46,4 @@ export function DocIframe({ URl, onConfirm }: Props) {
             </>
         }
     </>
-}
-
-const Sceleton = () => {
-    return <div className="shadow rounded-md w-full mx-auto px-2 md:px-4">
-        <div className="animate-pulse flex space-x-4">
-            <div className="flex-1 items-center space-y-6 py-1 content-start">
-                <div className="h-4 mx-auto w-1/2 place-self-center justify-self-center self-center bg-slate-700 rounded mb-4"></div>
-                <div className="space-y-6">
-                    {[...Array(6)]?.map((item, index) =>
-                        <div
-                            key={index}
-                        >
-                            <div className="h-2 bg-slate-700 rounded"></div>
-
-                            <div className="grid grid-cols-3 gap-4">
-                                <div className="h-2 bg-slate-700 rounded col-span-2"></div>
-                                <div className="h-2 bg-slate-700 rounded col-span-1"></div>
-                            </div>
-                        </div>
-                    )}
-                    <div className="h-2 bg-slate-700 rounded"></div>
-                </div>
-            </div>
-        </div>
-    </div>
 }
