+++ components/exchangesComponent.tsx
@@ -16,9 +16,10 @@ import { useAuthState } from "../context/authContext";
 import toast from "react-hot-toast";
 import shortenAddress, { shortenEmail } from "./utils/ShortenAddress";
 import HoverTooltip from "./Tooltips/HoverTooltip";
+import handleGoHome from "./utils/GoHome";
+import { ExchangesComponentSceleton } from "./Sceletons";
 
 interface UserExchange extends Exchange {
-
     note?: string,
     is_connected: boolean
 }
@@ -34,13 +35,6 @@ function UserExchanges() {
     const [exchangeLoading, setExchangeLoading] = useState<Exchange>()
     const { email } = useAuthState()
     const [exchangeToDisconnect, setExchangeToDisconnect] = useState<Exchange>()
-
-    const handleGoHome = useCallback(() => {
-        router.push({
-            pathname: "/",
-            query: router.query
-        })
-    }, [router.query])
 
     useEffect(() => {
 
@@ -192,7 +186,7 @@ function UserExchanges() {
                         </div>
                         <Combobox.Options static className="border-0 grid grid-cols-1 md:grid-cols-2 gap-2">
                             {
-                                loading ? <Sceleton />
+                                loading ? <ExchangesComponentSceleton />
                                     :
                                     <>
                                         {filteredItems?.length > 0 && (
@@ -333,29 +327,5 @@ function UserExchanges() {
         </div>
     )
 }
-
-const Sceleton = () => {
-
-    return <>
-        {[...Array(6)]?.map((item, index) =>
-            <div
-                key={index}
-                className="animate-pulse bg-darkblue-500 select-none rounded-lg p-3">
-                <div className="flex justify-between px-2">
-                    <div className="flex space-x-2">
-                        <div className="rounded-full bg-slate-700 h-8 w-8"></div>
-                        <div className="grid grid-cols-4">
-                            <div className="h-2 w-20 bg-slate-700 rounded col-span-3"></div>
-                        </div>
-                    </div>
-
-                    <div className="rounded bg-slate-700 h-8 w-20 place-self-end py-3 px-4"></div>
-                </div>
-            </div>
-        )}
-    </>
-
-}
-
 
 export default UserExchanges;
