+++ components/swapHistoryComponent.tsx
@@ -4,7 +4,6 @@ import LayerSwapApiClient, { SwapListResponse, SwapItem } from "../lib/layerSwap
 import TokenService from "../lib/TokenService"
 import SpinIcon from "./icons/spinIcon"
 import { ChevronRightIcon, ExternalLinkIcon, RefreshIcon, XIcon } from '@heroicons/react/outline';
-
 import { Dialog, Transition } from "@headlessui/react"
 import SwapDetails from "./swapDetailsComponent"
 import LayerswapMenu from "./LayerswapMenu"
@@ -16,7 +15,8 @@ import shortenAddress from "./utils/ShortenAddress"
 import { classNames } from "./utils/classNames"
 import SubmitButton from "./buttons/submitButton"
 import CopyButton from "./buttons/copyButton"
-
+import handleGoHome from "./utils/GoHome"
+import { SwapHistoryComponentSceleton } from "./Sceletons"
 
 export function StatusIcon({ swap }: { swap: SwapItem }) {
   if (swap?.status === 'failed') {
@@ -150,15 +150,6 @@ function TransactionsHistory() {
     setSelectedSwap(swap)
   }
 
-
-  const handleGoHome = useCallback(() => {
-    router.push({
-      pathname: "/",
-      query: router.query
-    })
-  }, [router.query])
-
-
   return (
     <div className={`bg-darkBlue px-8 md:px-12 shadow-card rounded-lg w-full overflow-hidden relative min-h`}>
       <div className="mt-3 flex items-center justify-between z-20" >
@@ -177,13 +168,12 @@ function TransactionsHistory() {
       </div>
       {
         page == 0 && loading ?
-          <Sceleton />
+          <SwapHistoryComponentSceleton />
           : <>
             {
               swaps?.data.length > 0 ?
                 <>
-                  <div className=" mb-2 ">
-
+                  <div className="mb-2">
                     <div className="-mx-4 mt-10 sm:-mx-6 md:mx-0 md:rounded-lg">
                       <table className="min-w-full divide-y divide-darkblue-100">
                         <thead>
@@ -491,7 +481,7 @@ function TransactionsHistory() {
                     </Dialog>
                   </Transition>
                 </>
-                : <div className="m-16 text-center mb-20 pb-10">
+                : <div className="sm:my-24 sm:mx-60 m-16 pb-20 text-center sm:pb-10">
                   There are no transactions for this account
                 </div>
             }
@@ -499,180 +489,6 @@ function TransactionsHistory() {
       }
     </div>
   )
-}
-
-const Sceleton = () => {
-
-  return <div className="animate-pulse">
-    <div className=" mb-10 ">
-      <div className="-mx-4 mt-10 sm:-mx-6 md:mx-0 md:rounded-lg ">
-        <table className="min-w-full divide-y divide-darkblue-100">
-          <thead>
-            <tr>
-              <th scope="col" className="py-3.5 pl-4 pr-3 text-left text-sm font-semibold text-gray-500 sm:pl-6">
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="hidden lg:block">
-                    <div className="h-2 w-8 bg-slate-700 rounded col-span-1"></div>
-                  </div>
-                </div>
-              </th>
-              <th
-                scope="col"
-                className="hidden px-3 py-3.5 text-left text-sm font-semibold text-gray-500 lg:table-cell"
-              >
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="h-2 w-8 bg-slate-700 rounded col-span-1"></div>
-                </div>
-              </th>
-              <th
-                scope="col"
-                className="px-3 py-3.5 text-left text-sm font-semibold text-gray-500 "
-              >
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="h-2 w-8 bg-slate-700 rounded col-span-1"></div>
-                </div>
-              </th>
-              <th
-                scope="col"
-                className="hidden px-3 py-3.5 text-left text-sm font-semibold text-gray-500 lg:table-cell"
-              >
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="h-2 w-16 bg-slate-700 rounded col-span-1"></div>
-                </div>
-              </th>
-              <th
-                scope="col"
-                className="hidden px-3 py-3.5 text-left text-sm font-semibold text-gray-500 lg:table-cell"
-              >
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="h-2 w-16 bg-slate-700 rounded col-span-1"></div>
-                </div>
-              </th>
-
-              <th
-                scope="col"
-                className="hidden px-3 py-3.5 text-left text-sm font-semibold text-gray-500 lg:table-cell"
-              >
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="h-2 w-8 bg-slate-700 rounded col-span-1"></div>
-                </div>
-              </th>
-              <th scope="col" className="relative py-3.5 pl-3 pr-4 sm:pr-6">
-                <div className="grid grid-cols-1 gap-4">
-                  <div className="h-2 w-8 bg-slate-700 rounded col-span-1"></div>
-                </div>
-              </th>
-              <th scope="col" className="relative py-3.5 pl-3 pr-4 sm:pr-6">
-
-              </th>
-            </tr>
-          </thead>
-          <tbody>
-            {[...Array(5)]?.map((item, index) => (
-              <tr key={index}>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'relative py-4 pl-4 sm:pl-6 pr-3 text-sm'
-                  )}
-                >
-                  <div className="text-white hidden lg:block">
-                    <div className="grid grid-cols-1 gap-4">
-                      <div className="h-2 w-16 bg-slate-700 rounded col-span-1"></div>
-                    </div>
-                  </div>
-                  {index !== 0 ? <div className="absolute right-0 left-6 -top-px h-px bg-darkblue-100" /> : null}
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'hidden px-3 py-3.5 text-sm text-white lg:table-cell'
-                  )}
-                >
-                  <div className="flex space-x-2">
-                    <div className="rounded-full bg-slate-700 h-4 w-4"></div>
-                    <div className="grid grid-cols-4 items-center">
-                      <div className="h-2 w-16 bg-slate-700 rounded col-span-3"></div>
-                    </div>
-                  </div>
-
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'px-3 py-3.5 text-sm text-white table-cell'
-                  )}
-                >
-                  <div className="flex space-x-2">
-                    <div className="rounded-full bg-slate-700 h-4 w-4"></div>
-                    <div className="grid grid-cols-4 items-center">
-                      <div className="h-2 w-16 bg-slate-700 rounded col-span-3"></div>
-                    </div>
-                  </div>
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'hidden px-3 py-3.5 text-sm text-white lg:table-cell'
-                  )}
-                >
-                  <div className="grid grid-cols-1 gap-4">
-                    <div className="h-2 w-16 bg-slate-700 rounded col-span-1"></div>
-                  </div>
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'relative px-3 py-3.5 text-sm text-white'
-                  )}
-                >
-                  <div className="grid grid-cols-1 gap-4">
-                    <div className="h-2 w-16 bg-slate-700 rounded col-span-1"></div>
-                  </div>
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'px-3 py-3.5 text-sm text-white  hidden lg:table-cell'
-                  )}
-                >
-                  <div className="flex space-x-2">
-                    <div className="rounded bg-slate-700 h-2 w-2"></div>
-                    <div className="grid grid-cols-1 items-center">
-                      <div className="h-2 w-16 bg-slate-700 rounded col-span-1"></div>
-                    </div>
-                  </div>
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'px-3 py-3.5 text-sm text-white  hidden lg:table-cell'
-                  )}
-                >
-                  <div className="grid grid-cols-2 gap-4">
-                    <div className="h-2 w-12 bg-slate-700 rounded col-span-1"></div>
-                    <div className="h-2 w-8 bg-slate-700 rounded col-span-1"></div>
-                  </div>
-                </td>
-                <td
-                  className={classNames(
-                    index === 0 ? '' : 'border-t border-darkblue-100',
-                    'px-3 py-3.5 text-sm text-white  hidden lg:table-cell'
-                  )}
-                >
-                  <div className="grid grid-cols-1 gap-4">
-                    <ChevronRightIcon className="h-5 w-5 text-slate-700" />
-                  </div>
-                </td>
-
-              </tr>
-            ))}
-          </tbody>
-        </table>
-      </div>
-    </div>
-  </div>
-
 }
 
 export default TransactionsHistory;
